--- a/sklearn/ensemble/_forest.py
+++ b/sklearn/ensemble/_forest.py
@@ -557,6 +557,21 @@ class BaseForest(MultiOutputMixin, BaseEnsemble, metaclass=ABCMeta):
             (accuracy score).
         """
 
+
+    def _get_oob_pred_parallel(self,
+                               estimator,
+                               n_samples,
+                               n_samples_bootstrap,
+                               X):
+        unsampled_indices = _generate_unsampled_indices(
+            estimator.random_state,
+            n_samples,
+            n_samples_bootstrap,
+        )
+        y_pred = self._get_oob_predictions(estimator, X[unsampled_indices, :])
+        return unsampled_indices, y_pred
+
+
     def _compute_oob_predictions(self, X, y):
         """Compute and set the OOB score.
 
@@ -597,14 +612,12 @@ class BaseForest(MultiOutputMixin, BaseEnsemble, metaclass=ABCMeta):
             n_samples,
             self.max_samples,
         )
-        for estimator in self.estimators_:
-            unsampled_indices = _generate_unsampled_indices(
-                estimator.random_state,
-                n_samples,
-                n_samples_bootstrap,
-            )
 
-            y_pred = self._get_oob_predictions(estimator, X[unsampled_indices, :])
+        all_pred = Parallel(n_jobs=self.n_jobs, verbose=self.verbose, require="sharedmem")(
+                delayed(self._get_oob_pred_parallel)(e, n_samples, n_samples_bootstrap, X)
+            for e in self.estimators_)
+
+        for unsampled_indices, y_pred in all_pred:
             oob_pred[unsampled_indices, ...] += y_pred
             n_oob_pred[unsampled_indices, :] += 1