--- a/sklearn/ensemble/_forest.py
+++ b/sklearn/ensemble/_forest.py
@@ -557,14 +557,24 @@ class BaseForest(MultiOutputMixin, BaseEnsemble, metaclass=ABCMeta):
             (accuracy score).
         """
 
-    def _get_oob_pred_parallel(self, estimator, n_samples, n_samples_bootstrap, X):
-        unsampled_indices = _generate_unsampled_indices(
-            estimator.random_state,
-            n_samples,
-            n_samples_bootstrap,
-        )
-        y_pred = self._get_oob_predictions(estimator, X[unsampled_indices, :])
-        return unsampled_indices, y_pred
+    def _get_oob_pred_parallel(
+        self, estimators, n_samples, n_samples_bootstrap, X, oob_pred, n_oob_pred, lock
+    ):
+        local_oob_pred = np.zeros_like(oob_pred, dtype=np.float32)
+        local_n_oob_pred = np.zeros_like(n_oob_pred, dtype=np.int8)
+        for estimator in estimators:
+            unsampled_indices = _generate_unsampled_indices(
+                estimator.random_state,
+                n_samples,
+                n_samples_bootstrap,
+            )
+            local_oob_pred[unsampled_indices, ...] += self._get_oob_predictions(
+                estimator, X[unsampled_indices, :]
+            )
+            local_n_oob_pred[unsampled_indices, :] += 1
+        with lock:
+            oob_pred += local_oob_pred
+            n_oob_pred += local_n_oob_pred
 
     def _compute_oob_predictions(self, X, y):
         """Compute and set the OOB score.
@@ -607,16 +617,18 @@ class BaseForest(MultiOutputMixin, BaseEnsemble, metaclass=ABCMeta):
             self.max_samples,
         )
 
-        all_pred = Parallel(
-            n_jobs=self.n_jobs, verbose=self.verbose, require="sharedmem", return_as="generator_unordered"
-        )(
-            delayed(self._get_oob_pred_parallel)(e, n_samples, n_samples_bootstrap, X)
-            for e in self.estimators_
-        )
+        lock = threading.Lock()
+        sub_ests = []
+        chunk_size = 40
+        for i in range(0, len(self.estimators_), chunk_size):
+            sub_ests.append(self.estimators_[i : i + chunk_size])
 
-        for unsampled_indices, y_pred in all_pred:
-            oob_pred[unsampled_indices, ...] += y_pred
-            n_oob_pred[unsampled_indices, :] += 1
+        Parallel(n_jobs=self.n_jobs, verbose=self.verbose, require="sharedmem")(
+            delayed(self._get_oob_pred_parallel)(
+                ests, n_samples, n_samples_bootstrap, X, oob_pred, n_oob_pred, lock
+            )
+            for ests in sub_ests
+        )
 
         for k in range(n_outputs):
             if (n_oob_pred == 0).any():