* Fixes #28059

* use `joblib` to make oob predictions in random forest

* I also drafted in some `asv` benchmarks to try to capture the performance improvements more formally, but it looks like the `virtualenv` machinery for running `asv` may need a tune up? I kept seeing `Failed: trying different commit/environment` via `asv continuous -E virtualenv -e -b "RandomForestClassifierBenchmark" main treddy_issue_28059` and that was after having to apply a patch to `asv.conf.json` for a missing dependency (perhaps the team more commonly uses `conda` here)

* in lieu of the more formal benchmarks working (it would be good to confirm those...) I re-ran the code from the matching issue and the scaling does look better now with 16 cores on that machine:

<details>

![bench_feat](https://github.com/scikit-learn/scikit-learn/assets/7903078/c5a2fef3-9467-406a-8dec-fc19ac0d6c5b)

</details>