@tylerjereddy 2024-07-10
Oh right, I know Thomas. I guess if I were to criticize this I might wonder if some of the other tree-based methods might benefit (is the infrastructure general enough?), but maybe that could be done later.

Did we get the benchmarks tested? It looks I didn't actually sort out the `asv` run, but should probably confirm that works.