@tylerjereddy 2024-07-10
One other thing to confirm might be that the tests properly flush the new parallel code path? Does that happen automatically since `n_jobs` is already used productively elsewhere?