@adam2392 2024-07-10
> Oh right, I know Thomas. I guess if I were to criticize this I might wonder if some of the other tree-based methods might benefit (is the infrastructure general enough?), but maybe that could be done later.

I think the only other place I can currently think of is in `_bagging.py`, but that is a separate file and the implementation will slightly differ I imagine. I feel like that can be done later, but perhaps someone else disagrees.

> Did we get the benchmarks tested? It looks I didn't actually sort out the `asv` run, but should probably confirm that works.

Yes that would be nice to run. Would you be able to run the asv benchmark you proposed, or run a similar setting as the one you posted in the [GH issue](https://github.com/scikit-learn/scikit-learn/issues/28059#issue-2064663387), where you also vary the # of samples?