@tylerjereddy 2024-09-06
For:

`export SKLBENCH_NJOBS=[1,8]`
`asv continuous -E virtualenv -e -b "RandomForestClassifierBenchmark" main treddy_issue_28059`

There is one benchmark that gets faster, `oob` active with concurrency as you might expect, but also some memory cost/overhead associated:

```
| Change   | Before [0b0b90b7] <main>   | After [44d6e409] <treddy_issue_28059>   |   Ratio | Benchmark (Parameter)                                                   |
|----------|----------------------------|-----------------------------------------|---------|-------------------------------------------------------------------------|
| +        | 211M                       | 266M                                    |    1.26 | ensemble.RandomForestClassifierBenchmark.peakmem_fit('dense', 8, True)  |
| +        | 434M                       | 521M                                    |    1.2  | ensemble.RandomForestClassifierBenchmark.peakmem_fit('sparse', 8, True) |
| +        | 210M                       | 246M                                    |    1.17 | ensemble.RandomForestClassifierBenchmark.peakmem_fit('dense', 1, True)  |
| -        | 1.22±0.06s                 | 916±10ms                                |    0.75 | ensemble.RandomForestClassifierBenchmark.time_fit('sparse', 8, True)    |


```

Using the larger scale benchmarks shows a more dramatic performance improvement starting to show up. There is still some memory cost.

`export SKLBENCH_PROFILE=large_scale`
`export SKLBENCH_NJOBS=[16]`
`asv continuous -E virtualenv -e -b "RandomForestClassifierBenchmark" main treddy_issue_28059`

```
| Change   | Before [0b0b90b7] <main>   | After [44d6e409] <treddy_issue_28059>   |   Ratio | Benchmark (Parameter)                                                   |
|----------|----------------------------|-----------------------------------------|---------|-------------------------------------------------------------------------|
| +        | 601M                       | 789M                                    |    1.31 | ensemble.RandomForestClassifierBenchmark.peakmem_fit('dense', 16, True) |
| -        | 4.47±0.08s                 | 2.69±0.07s                              |    0.6  | ensemble.RandomForestClassifierBenchmark.time_fit('sparse', 16, True)   |
```