@tylerjereddy 2024-09-09
Naive addition of `return_as="generator"` seemed to nullify the performance gains; I tried the fancy new `generator_unordered` as well, but seemed the same. Maybe I need to think about the design a bit more carefully in the generator context though.

<details>

```diff
--- a/sklearn/ensemble/_forest.py
+++ b/sklearn/ensemble/_forest.py
@@ -608,7 +608,7 @@ class BaseForest(MultiOutputMixin, BaseEnsemble, metaclass=ABCMeta):
         )
 
         all_pred = Parallel(
-            n_jobs=self.n_jobs, verbose=self.verbose, require="sharedmem", return_as="generator"
+            n_jobs=self.n_jobs, verbose=self.verbose, require="sharedmem", return_as="generator_unordered"
         )(
             delayed(self._get_oob_pred_parallel)(e, n_samples, n_samples_bootstrap, X)
             for e in self.estimators_

```


</details>